+++ src/dbn_file_store.cpp
@@ -12,13 +12,12 @@ DbnFileStore::DbnFileStore(const std::string& file_path)
 void DbnFileStore::Replay(const MetadataCallback& metadata_callback,
                           const RecordCallback& record_callback) {
   auto metadata = parser_.DecodeMetadata();
-  const auto record_count = metadata.record_count;
   if (metadata_callback) {
     metadata_callback(std::move(metadata));
   }
-  for (std::size_t i = 0; i < record_count; ++i) {
-    const auto record = parser_.DecodeRecord();
-    if (record_callback(record) == KeepGoing::Stop) {
+  const databento::Record* record;
+  while ((record = parser_.DecodeRecord()) != nullptr) {
+    if (record_callback(*record) == KeepGoing::Stop) {
       break;
     }
   }